Former-commit-id: 17e9f7606544c101b2d20cea21ce3a3458f14555
Former-commit-id: 9df696a78cefb3a0a2071a67df0d0d9df79fe7cf
Former-commit-id: 80dff4f3c5542d8b8f70a2e204752932eccfcf55

+++ src/main/java/teammates/common/util/TimeHelper.java
@@ -18,7 +18,7 @@ public class TimeHelper {
 	/**
 	 * Returns the current date and time as a {@code Calendar} object for the given timezone.
 	 */
-	public static Calendar now(int timeZone) {
+	public static Calendar now(double timeZone) {
 		return TimeHelper.convertToUserTimeZone(
 				Calendar.getInstance(TimeZone.getTimeZone("UTC")), timeZone);
 	}
